Refactored some event reading and plotting
